+++ solawi-bid-backend/src/main/kotlin/org/solyton/solawi/bid/application/pipeline/validate-headers.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.json.Json
 import org.evoleq.ktorx.result.Result
 import org.evoleq.ktorx.result.ResultSerializer
 import org.solyton.solawi.bid.application.exception.ApplicationException
-import org.solyton.solawi.bid.application.permission.Context
+import org.solyton.solawi.bid.module.application.permission.Context
 import org.solyton.solawi.bid.application.permission.Header
 
 fun Application.interceptAndValidateHeaders() {
